[기초관리] - [상품관리] - [키오스크키맵관리] - 중분류 사용하던 키맵 삭제후 키맵매장적용 시, 중분류사용여부 체크 부분 수정 (중분류사용여부 값이 없으면 대분류사용으로 처리) - 중분류 사용하는 경우, 중분류를 추가/삭제 하는 과정에서 대분류코드 값이 잘못 들어가는 오류 수정 - [키오스크 키맵복사 탭] 에서 키맵복사 후, 기준매장 리스트 재조회

+++ src/main/java/kr/co/solbipos/base/prod/kioskKeyMap/service/impl/KioskKeyMapServiceImpl.java
@@ -565,7 +565,11 @@ public class KioskKeyMapServiceImpl implements KioskKeyMapService {
 
                 // 해당 키맵그룹의 중분류 사용여부 조회
                 List<DefaultMap<Object>> list = kioskKeyMapMapper.getKioskKeyMapGroupTuMClsFg(kioskKeyMapVO);
-                kioskKeyMapVO.setTuMClsFg(list.get(0).getStr("tuMClsFg"));
+                if(list.size() > 0){
+                    kioskKeyMapVO.setTuMClsFg(list.get(0).getStr("tuMClsFg"));
+                }else{
+                    kioskKeyMapVO.setTuMClsFg("0");
+                }
 
                 // 새 키맵그룹과 카테고리(분류)코드로 INSERT
                 result = kioskKeyMapMapper.mergeKioskCategoryStoreReg(kioskKeyMapVO);
@@ -722,7 +726,11 @@ public class KioskKeyMapServiceImpl implements KioskKeyMapService {
 
                 // 해당 키맵그룹의 중분류 사용여부 조회
                 List<DefaultMap<Object>> list = kioskKeyMapMapper.getKioskKeyMapGroupTuMClsFg(kioskKeyMapVO);
-                kioskKeyMapVO.setTuMClsFg(list.get(0).getStr("tuMClsFg"));
+                if(list.size() > 0){
+                    kioskKeyMapVO.setTuMClsFg(list.get(0).getStr("tuMClsFg"));
+                }else{
+                    kioskKeyMapVO.setTuMClsFg("0");
+                }
 
                 // 새 키맵그룹과 카테고리(분류)코드로 INSERT
                 result = kioskKeyMapMapper.mergeKioskCategoryStoreReg(kioskKeyMapVO);
@@ -1275,9 +1283,11 @@ public class KioskKeyMapServiceImpl implements KioskKeyMapService {
 
             kioskKeyMapClassVO.setOrgTuClsType(kioskKeyMapVO.getTuClsType());
             List<DefaultMap<Object>> selectList = kioskKeyMapMapper.getKioskMClsCount(kioskKeyMapClassVO);
-            if (Integer.parseInt(selectList.get(0).get("rowCount").toString()) > 0) {
-               // 키오스크 중분류 삭제
-               kioskKeyMapMapper.deleteAllKioskKeyMClass(kioskKeyMapClassVO);
+            if(selectList.size() > 0) {
+                if (Integer.parseInt(selectList.get(0).get("rowCount").toString()) > 0) {
+                    // 키오스크 중분류 삭제
+                    kioskKeyMapMapper.deleteAllKioskKeyMClass(kioskKeyMapClassVO);
+                }
             }
 
             kioskKeyMapVO.setOrgnFg(sessionInfoVO.getOrgnFg().getCode());
